build: Drop distutils from setup.py
* setup.py.in: Drop fallback code for distutils in case setuptools is
not available
--

We are now using the pyproject.toml file to statically declare our build
time dependencies.  This ensures that a compatible version of setuptools
should be available by the time the setup.py file is executed.  For that
reason we do not need the fallback code any longer.

Also order imports according to PEP-8 and drop unused "glob" import.

Signed-off-by: Lucas Hoffmann <lucc@posteo.de>